Exception in package Business

+++ src/business/CollectiveManager.java
@@ -2,6 +2,7 @@ package business;
 
 import dataAcces.CollectiveDB;
 import dataAcces.dao.ICollectiveDAO;
+import exception.BusinessException;
 import exception.DAOException;
 import exception.ModelException;
 import model.Collective;
@@ -33,11 +34,12 @@ public class CollectiveManager {
     }
 
     // Delete a collective
-    public void deleteCollective(int idCollective) throws DAOException {
+    public void deleteCollective(int idCollective) throws DAOException, BusinessException {
+        if (idCollective < 1) throw new BusinessException("Erreur : l'identifiant doit être suppérieur à 0");
         dao.delete(idCollective);
     }
 
-    public void deleteCollective(Collective collective) throws DAOException {
+    public void deleteCollective(Collective collective) throws DAOException, BusinessException {
         this.deleteCollective(collective.getIdCollective());
     }
 
@@ -47,12 +49,14 @@ public class CollectiveManager {
     }
 
     // Search for a collective based on its id
-    public Collective getACollectiveBasedId(int idCollective) throws DAOException, ModelException {
+    public Collective getACollectiveBasedId(int idCollective) throws DAOException, ModelException, BusinessException {
+        if (idCollective < 1) throw new BusinessException("Erreur : l'identifiant doit être suppérieur à 0");
         return dao.getACollectiveBasedId(idCollective);
     }
 
     // Retrieving a collective's ID based on its name
-    public int getACollectiveIDBasedName(String name) throws DAOException, ModelException {
+    public int getACollectiveIDBasedName(String name) throws DAOException, ModelException, BusinessException {
+        if (name.equals("")) throw new BusinessException("Erreur : le nom ne peut être une chaine vide");
         return dao.getACollectiveIDBasedName(name);
     }
