+++ src/business/OneObjectManager.java
@@ -2,6 +2,7 @@ package business;
 
 import dataAcces.OneObjectDB;
 import dataAcces.dao.IOneObjectDAO;
+import exception.BusinessException;
 import exception.ModelException;
 import model.Collective;
 import model.OneObject;
@@ -34,11 +35,12 @@ public class OneObjectManager {
     }
 
     // Delete an object
-    public void deleteAnObject(int idObject) throws DAOException {
+    public void deleteAnObject(int idObject) throws DAOException, BusinessException {
+        if (idObject < 1) throw new BusinessException("Erreur : l'identifiant doit être suppérieur à 0");
         dao.delete(idObject);
     }
 
-    public void deleteAnObject(OneObject oneObject) throws DAOException {
+    public void deleteAnObject(OneObject oneObject) throws DAOException, BusinessException {
         this.deleteAnObject(oneObject.getIdObject());
     }
 
@@ -48,16 +50,18 @@ public class OneObjectManager {
     }
 
     // Recovering all the objects of a collective
-    public ArrayList<OneObject> getAllObjectsForOneCollective(int idCollective) throws DAOException, ModelException {
+    public ArrayList<OneObject> getAllObjectsForOneCollective(int idCollective) throws DAOException, ModelException, BusinessException {
+        if (idCollective < 1) throw new BusinessException("Erreur : l'identifiant doit être suppérieur à 0");
         return dao.getAllObjectsForOneCollective(idCollective);
     }
 
-    ArrayList<OneObject> getAllObjectsForOneCollective(Collective collective) throws DAOException, ModelException {
+    ArrayList<OneObject> getAllObjectsForOneCollective(Collective collective) throws DAOException, ModelException, BusinessException {
         return this.getAllObjectsForOneCollective(collective.getIdCollective());
     }
 
     // Retrieving an object via its id
-    public OneObject getObjectByID(int idObject) throws DAOException, ModelException {
+    public OneObject getObjectByID(int idObject) throws DAOException, ModelException, BusinessException {
+        if (idObject < 1) throw new BusinessException("Erreur : l'identifiant doit être suppérieur à 0");
         return dao.getObjectByID(idObject);
     }
 }
